requirements.txt
openai
python-dotenv
notion-client
python-telegram-bot
flask
